change double input values to satoshis

+++ libraries/plugins/peerplays_sidechain/include/graphene/peerplays_sidechain/sidechain_net_handler_bitcoin.hpp
@@ -20,7 +20,7 @@ class btc_txout {
 public:
    std::string txid_;
    unsigned int out_num_;
-   double amount_;
+   uint64_t amount_;
 };
 
 class bitcoin_rpc_client {
